+++ spec/paper_trail_spec.rb
@@ -52,23 +52,4 @@ RSpec.describe PaperTrail do
     end
   end
 
-  describe "deprecated methods" do
-    before do
-      allow(ActiveSupport::Deprecation).to receive(:warn)
-    end
-
-    shared_examples "it delegates to request" do |method, args|
-      it do
-        #arguments = args || [no_args]
-        #allow(described_class.request).to receive(method)
-        described_class.public_send(method, *args)
-        #expect(described_class.request).to have_received(method).with(*arguments)
-        expect(ActiveSupport::Deprecation).to have_received(:warn)
-      end
-    end
-
-    it_behaves_like "it delegates to request", :clear_transaction_id, nil
-    it_behaves_like "it delegates to request", :transaction_id=, 123
-    it_behaves_like "it delegates to request", :transaction_id, nil
-  end
 end